feat: make valid response for unauthorized calls,  make user endpoints accesible only for admin user

+++ src/main/java/kg/edu/iaau/beeline/controller/PersonController.java
@@ -6,6 +6,7 @@ import kg.edu.iaau.beeline.other.CustomResponse;
 import kg.edu.iaau.beeline.service.PersonService;
 import kg.edu.iaau.beeline.transfer.Groups;
 import kg.edu.iaau.beeline.transfer.View;
+import kg.edu.iaau.beeline.util.ResponseUtil;
 import org.modelmapper.Conditions;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +17,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
@@ -36,41 +38,59 @@ public class PersonController
     @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
 
+    @Autowired
+    private ResponseUtil responseUtil;
+
     @GetMapping
-    public ResponseEntity<CustomResponse> getPosts()
+    public ResponseEntity getPosts(Principal principal)
     {
+        if(!personService.isAdmin(principal.getName()))
+        {
+            return responseUtil.responseBuilder("ERROR",
+                    "notAllowed", HttpStatus.METHOD_NOT_ALLOWED);
+        }
+
         List<Person> personList = personService.getAll();
         List<PersonDTO> personDTOList = personList.stream()
                 .map(person -> mapper.map(person, PersonDTO.class))
                 .collect(Collectors.toList());
 
-        return new ResponseEntity(
-                new CustomResponse("SUCCESS",
-                        messageSource.getMessage("getUser",new Object[0], new Locale("")),
-                        personDTOList), HttpStatus.OK);
+        return responseUtil.responseBuilder("SUCCESS",
+                "getUser", personDTOList, HttpStatus.OK);
     }
 
     @PostMapping
-    public ResponseEntity<PersonDTO> create(@Validated(Groups.New.class)
+    public ResponseEntity create(Principal principal, @Validated(Groups.New.class)
                                             @RequestBody PersonDTO personDTO)
     {
+        if(!personService.isAdmin(principal.getName()))
+        {
+            return responseUtil.responseBuilder("ERROR",
+                    "notAllowed", HttpStatus.METHOD_NOT_ALLOWED);
+        }
+
         Person person = mapper.map(personDTO, Person.class);
         person.setPassword(bCryptPasswordEncoder.encode(person.getPassword()));
 
         Person personCreated = personService.save(person);
         PersonDTO dto = mapper.map(personCreated, PersonDTO.class);
 
-        return new ResponseEntity(
-                new CustomResponse("SUCCESS",
-                        messageSource.getMessage("postUser",new Object[0], new Locale("")),
-                        dto), HttpStatus.OK);
+        return responseUtil.responseBuilder("SUCCESS",
+                "postUser", dto, HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity updatePerson(@PathVariable int id,
-                                  @Validated(Groups.Update.class)
-                                  @RequestBody PersonDTO personDTO)
+    public ResponseEntity updatePerson( Principal principal,
+                                        @PathVariable int id,
+                                        @Validated(Groups.Update.class)
+                                        @RequestBody PersonDTO personDTO)
     {
+        if(!personService.isAdmin(principal.getName()))
+        {
+            return responseUtil.responseBuilder("ERROR",
+                    "notAllowed", HttpStatus.METHOD_NOT_ALLOWED);
+        }
+
         Person person = personService.getById(id);
         mapper.getConfiguration().setPropertyCondition(Conditions.isNotNull());
         mapper.map(personDTO,person);
@@ -79,21 +99,23 @@ public class PersonController
 
         personDTO = mapper.map(person, PersonDTO.class);
 
-        return new ResponseEntity(
-                new CustomResponse("SUCCESS",
-                        messageSource.getMessage("infoUpdated",new Object[0], new Locale("")),
-                        personDTO), HttpStatus.OK);
+        return responseUtil.responseBuilder("SUCCESS",
+                "infoUpdated", personDTO , HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deletePerson(@PathVariable int id)
+    public ResponseEntity deletePerson(@PathVariable int id, Principal principal)
     {
+        if(!personService.isAdmin(principal.getName()))
+        {
+            return responseUtil.responseBuilder("ERROR",
+                    "notAllowed", HttpStatus.METHOD_NOT_ALLOWED);
+        }
+
         Person person = personService.getById(id);
         personService.delete(person);
 
-        return new ResponseEntity(
-                new CustomResponse("SUCCESS",
-                        messageSource.getMessage("deleteUser",new Object[0], new Locale(""))),
-                        HttpStatus.OK);
+        return responseUtil.responseBuilder("SUCCESS",
+                "deleteUser", HttpStatus.OK);
     }
 }
